+++ csv_results/summary_stat_by_cluster.csv
@@ -1,0 +1,6 @@
+"","cluster","mean_time","mean_time_mod","mean_pass","mean_pass_mod","mean_ozone_time","mean_max_xG","mean_cmltv_xG","mean_goals","mean_shots","mean_passes","mean_slotpasses","mean_fwdtime","mean_dmentime","mean_penalties","mean_puckretrievals","mean_icings"
+"1",1,8.47540390196079,10.5561264537275,2.15479876160991,2.68380462724936,6.516029336,0.040745381,0.041543812,0.052069426,1.113484646,1.560747664,0.238985314,3.813181029,1.896893834,0.005340454,0.893190921,0.009345794
+"2",2,9.17301749785592,11.0621186953004,2.42457614673628,2.9234984983082,7.0240521652,0.049428114,0.0508152946,0.0483171096,1.1224919594,1.8301682088,0.3097903604,4.212393564,1.8761012356,0.0079844056,0.9521138568,0.0067401094
+"3",3,9.18678607983516,11.3587271549501,2.26430253454794,2.79864017644709,6.576771011,0.0518686315,0.053324156,0.0536633315,1.1014132495,1.578616498,0.2711678605,3.7668253415,1.895598095,0.013105752,0.8762583375,0.010509305
+"4",4,10.0722199244103,12.0824608246567,2.68121500338289,3.21667036151572,7.80779308766667,0.05064824,0.0525784076666667,0.041718824,1.13055755166667,1.984277373,0.316885631333333,4.51772296633333,2.30768356133333,0.00886516133333333,0.964308609333333,0.0138518636666667
+"5",5,10.1331234352247,11.9963663160868,2.72396285649831,3.22454444988635,7.384858041,0.0448349163333333,0.0457715446666667,0.0591676473333333,1.097496845,1.962949011,0.254945748666667,4.177584168,2.262192677,0.00562638933333333,0.928281403,0.00794342166666667
