--- data_clustered.csv
@@ -1,15 +1,0 @@
-"","team","mean_time","mean_time_mod","mean_pass","mean_pass_mod","mean_ozone_time","mean_max_xG","mean_cmltv_xG","mean_goals","mean_shots","mean_passes","mean_slotpasses","mean_fwdtime","mean_dmentime","mean_penalties","mean_puckretrievals","mean_icings","cluster"
-"1",795,8.47540390196079,10.5561264537275,2.15479876160991,2.68380462724936,6.516029336,0.040745381,0.041543812,0.052069426,1.113484646,1.560747664,0.238985314,3.813181029,1.896893834,0.005340454,0.893190921,0.009345794,1
-"2",814,8.89264354745925,10.7974705704307,2.30968360498562,2.80442374854482,6.798044304,0.050211086,0.051776869,0.047029703,1.129950495,1.726485149,0.308168317,4.053143184,1.755669035,0.006188119,0.917079208,0.002475248,2
-"3",877,8.93431896180215,10.8983747431302,2.41821743388834,2.94982078853047,6.98779542,0.04599254,0.047365598,0.042606516,1.130325815,1.812030075,0.253132832,4.229791516,1.835446793,0.00877193,0.93358396,0.00877193,2
-"4",885,9.15308717280163,11.4618684443022,2.19325153374233,2.74647887323944,6.726561614,0.05097762,0.052191131,0.04743083,1.085638999,1.528326746,0.279314888,3.892157067,1.851768578,0.014492754,0.881422925,0.015810277,3
-"5",524,9.22048498686869,11.255585865598,2.33535353535354,2.85080147965475,6.426980408,0.052759643,0.054457181,0.059895833,1.1171875,1.62890625,0.263020833,3.641493616,1.939427612,0.01171875,0.87109375,0.005208333,3
-"6",869,9.27517381182216,11.2681863664286,2.31681243926142,2.814639905549,7.280018338,0.049893469,0.051218329,0.054568528,1.13071066,1.793147208,0.312182741,4.388553635,1.87859686,0.01142132,0.997461929,0.008883249,2
-"7",855,9.29632114122733,11.1389423392967,2.55389541088581,3.06010230179028,7.133486691,0.051696839,0.05301402,0.063087248,1.096644295,1.926174497,0.348993289,4.206774626,2.025781988,0.008053691,0.989261745,0.006711409,2
-"8",792,9.46663002696872,11.2076194572158,2.52427184466019,2.98850574712644,6.920916073,0.049346636,0.050701657,0.034293553,1.124828532,1.893004115,0.326474623,4.183704859,1.885011502,0.005486968,0.923182442,0.006858711,2
-"9",726,9.86811964622642,11.7004550615213,2.64245283018868,3.1331096196868,7.541143872,0.045032984,0.047177381,0.033613445,1.12244898,1.837935174,0.306122449,4.31879089,2.144480929,0.010804322,0.948379352,0.013205282,4
-"10",628,9.99651806156716,11.7502931600877,2.73507462686567,3.21491228070175,7.047171342,0.042969869,0.043895441,0.054373522,1.113475177,1.93498818,0.258865248,3.911930599,2.146163045,0.007092199,0.90070922,0.007092199,5
-"11",634,10.1222102794457,11.8939404369064,2.70785219399538,3.18181818181818,7.307320583,0.044375592,0.045256725,0.060029283,1.0761347,1.920937042,0.257686676,4.201297653,2.17367488,0.002928258,0.859443631,0.004392387,5
-"12",686,10.1280039650901,12.336992484225,2.74436936936937,3.34293552812071,8.069190586,0.052946714,0.05495397,0.041543027,1.12462908,2.197329377,0.317507418,4.757459033,2.368922004,0.010385757,0.954005935,0.014836795,4
-"13",503,10.2205361619145,12.2099349282238,2.65682281059063,3.17396593673966,7.813044805,0.053965022,0.055603872,0.05,1.144594595,1.917567568,0.327027027,4.476918976,2.409647751,0.005405405,0.990540541,0.013513514,4
-"14",825,10.2806419646612,12.3448653512664,2.72896174863388,3.27690288713911,7.800082198,0.047159288,0.048162468,0.063100137,1.102880658,2.032921811,0.248285322,4.419524252,2.466740106,0.006858711,1.024691358,0.012345679,5
