+++ final_scripts/LinHac.Rmd
@@ -1,0 +1,111 @@
+---
+title: "LinHac"
+author: "Xikai wu"
+date: "2024-04-25"
+output: html_document
+---
+
+```{r setup, include=FALSE}
+library(tidyverse)
+library(knitr)
+library(ggplot2)
+```
+
+```{r}
+df=read.csv("Linhac24_Sportlogiq.csv")
+```
+
+```{r}
+df_pass_between_shots=df %>% group_by(gameid) %>%
+  mutate(current_shot=cumsum(eventname=="shot"))%>%
+  #filter(eventname=="pass")#%>%
+  #group_by(gameid,current_shot)%>%
+  filter(eventname=="pass")%>%
+  add_count(gameid,current_shot,name="pass_between")
+df_pass_by_team=df%>% group_by(gameid)%>%
+  mutate(current_shot=cumsum(eventname=="shot"))%>%
+  filter(eventname=="pass")%>%
+  add_count(gameid,current_shot,teamid,name="pass_between")
+df_joined_passbetween=left_join(df,df_pass_between_shots)
+```
+
+```{r}
+teams=unique(df$teamid)
+#for (i in teams){
+#  df_pass_by_team%>%filter(teamid==i)%>%group_by(gameid,current_shot)
+#}
+temp=df_pass_by_team%>%filter(teamid==814)%>%group_by(gameid,current_shot)%>%summarize(pass_count=pass_between[1],teamid=teamid[1])
+
+
+
+## Try to figure out passes leading up to a shot
+df_possession_id=df%>%filter(!is.na(teaminpossession))%>%#group_by(gameid)%>% add group by gameid for per game
+  mutate(possession_id=cumsum(c(1, diff(teaminpossession) != 0)))
+## if a possession contain a shot attempt, then count number of passes within the posession that leads up to the shot
+containShot=function(event){
+  has_shot=(event=="shot")
+  if (sum(has_shot)>0){
+    return(rep(TRUE,length(event)))
+  }
+  else{
+    return(rep(FALSE,length(event)))
+  }
+}
+posession_with_shot=df_possession_id%>%group_by(possession_id)%>%reframe(has_shot=containShot(eventname))
+df_possession_id$has_shot=posession_with_shot$has_shot
+##now i have df with posession column and filter by whether the possession contains a shot
+## then we could groupby possession then count the pass between each shot.
+df_pass_betweenshot_per_possession=df_possession_id%>%filter(has_shot==T)%>%group_by(possession_id)%>%mutate(pass_count=cumsum(eventname=="pass"),time_since_possession=compiledgametime-compiledgametime[1])%>%filter(eventname=="shot")%>%mutate(passes_between_shots = pass_count - lag(pass_count, default =0),time_since_shot=time_since_possession-lag(time_since_possession,default=0))
+
+
+#cleaned_data
+df_pass_shot_all=df_pass_betweenshot_per_possession%>%select(gameid,teaminpossession,possession_id,eventname,outcome,pass_count,passes_between_shots,time_since_possession,time_since_shot,manpowersituation)%>%mutate(teaminpossession=factor(teaminpossession))
+df_pass_shot_clean=df_pass_shot_all[df_pass_shot_all$manpowersituation=="evenStrength",]
+#write.csv(df_pass_shot_clean, "clean_df_with_pass_time_data.csv", row.names = FALSE)
+#write.csv(df_pass_betweenshot_per_possession, "full_df_with_pass_time_data.csv", row.names = FALSE)
+
+#df_pass_shot_clean%>%group_by(teaminpossession)
+
+### Use GG plot and graph histogram for each team
+
+#shot_time_per_team=ggplot(df_pass_shot_clean,aes(x=time_since_shot,fill=teaminpossession))+
+#geom_histogram(alpha=0.1,position="identity")
+#shot_time_per_team
+#ggsave("ShotTimeHistogram.jpg",shot_time_per_team, 
+#        dpi = 300, 
+#        width = 5, #long width
+ #       height = 9) #short height
+
+summary_shot_pass=df_pass_shot_clean%>%group_by(teaminpossession)%>%summarize(teaminpossession=teaminpossession[1],
+                                                            mean_shot_interval=mean(time_since_shot),median_shot_interval=median(time_since_shot),
+                                                            mean_pass_per_shot=mean(passes_between_shots),median_pass_per_shot=median(passes_between_shots))%>%
+  arrange(mean_shot_interval)
+#kable(summary_shot_pass)
+
+### want a more robust measure that takes into account that multiple shots will be taken in quick succession so that the mean shot interval could be 
+
+interval_counts <- df_pass_shot_clean %>%
+  group_by(possession_id) %>%
+  mutate(interval = cut(time_since_possession, breaks = seq(0, max(time_since_possession) + 0.5, by = 0.5))) %>%
+  summarise(count_interval = length(unique(interval)))
+#not as useful as time between shots could even be more than a second if even if theres no pass between
+df_withinterval=left_join(df_pass_shot_clean,interval_counts)
+
+## instead just count the number of shots with 0 passes in between
+
+pass_count_non_zero=df_pass_shot_clean%>%group_by(teaminpossession)%>%summarize(count_with_pass=sum(passes_between_shots!=0))
+df_passshot_with_count=left_join(df_pass_shot_clean,pass_count_non_zero,by = join_by(teaminpossession))
+
+df_modified_total_shot=df_passshot_with_count%>%group_by(teaminpossession)%>%summarize(teaminpossession=teaminpossession[1],
+                                                            mean_shot_interval_mod=sum(time_since_shot)/count_with_pass[1],median_shot_interval=median(time_since_shot),
+                                                            mean_pass_per_shot=sum(passes_between_shots)/count_with_pass[1],median_pass_per_shot=median(passes_between_shots),count=count_with_pass[1])%>%
+  arrange(mean_shot_interval_mod)
+
+df_rank=full_join(mutate(select(df_modified_total_shot,teaminpossession,mean_shot_interval_mod),rank_modified=rank(mean_shot_interval_mod)),mutate(select(summary_shot_pass,teaminpossession,mean_shot_interval),rank=rank(mean_shot_interval)),by = join_by(teaminpossession))
+
+colnames(df_modified_total_shot)=c("teaminpossession","mean_time_mod","median_time_mod","mean_pass_mod","median_pass_mod","zero_pass_shot_count")
+colnames(summary_shot_pass)=c("teaminpossession","mean_time","median_time","mean_pass","median_pass")
+df_final=full_join(summary_shot_pass,df_modified_total_shot)
+
+#write.csv(df_final, "var_for_clustering.csv", row.names = FALSE)
+```
